pyproject.toml
[project]
name = "discord-storage-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ecdsa>=0.19.0",
    "eciespy>=0.4.3",
    "fire>=0.7.0",
]
